pyproject.toml
[project]
name = "anmoku"
description = "A peaceful and fully typed MyAnimeList/Jikan API wrapper with caching and proper rate limiting."
authors = [
    {name = "Goldy", email = "goldy@devgoldy.xyz"},
    {name = "EmreTech", email = "emreterzioglu49@gmail.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = [
    "anime api wrapper", "mal", "myanimelist", "My Anime List", "jikan api wrapper", "async anime api wrapper", "jikan caching", "jikan rate limiting"
]
classifiers = [
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: POSIX :: Linux",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "importlib-metadata; python_version<'3.8'",
    "typing-extensions",
    "devgoldyutils>=2.5.6",
    "requests",
    "aiohttp",
    "Pillow",
    "slowstack"
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ruff",
    "build",
    "pytest",
    "pytest-asyncio",
    "Sphinx",
    "sphinxext-opengraph",
    "sphinx-inline-tabs",
    "furo",
    "myst-parser",
    "sphinx-copybutton"
]

[project.urls]
GitHub = "https://github.com/THEGOLDENPRO/anmoku"
BugTracker = "https://github.com/THEGOLDENPRO/anmoku/issues"

[tool.setuptools.dynamic]
version = { attr = "anmoku.__version__" }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["anmoku*"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]